requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
Pillow==9.1.1
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
